refactor: move RecordScorer from services to models (domain layer)
RecordScorer is a pure domain calculator — no orchestration, no side effects.
It computes scoring for vintage, condition, desirability, freshness, and noise.
Moving it to app/models/ places it alongside StorefrontTheme as a non-AR
domain object, resolving the primary layer violation from the layered
architecture analysis.
