+++ app/models/record_scorer.rb
